+++ src/controllers/item.controller.js
@@ -4,12 +4,18 @@
 'use strict';
 
 import { Item } from '../models/data/item.model.js';
+import { MessageQueue } from '../models/data/message-queue.model.js';
+import { AmqpHandler } from '../services/amqp-handler.service.js';
 
 
-export class ItemController {
+export class ItemController extends AmqpHandler {
 
   constructor(itemsenseService) {
+    const { username, password } = itemsenseService._itemsenseConfig;
+    super(username, password);
+
     this.model = new Item();
+    this.queueModel = new MessageQueue();
     this.itemsenseService = itemsenseService;
   }
 
@@ -19,6 +25,10 @@ export class ItemController {
 
   getHistory(queryParams) {
     return this.itemsenseService.makeRequest(this.model, Item.requestTypes.HISTORY, null, null, queryParams);
+  }
+
+  configureQueue(messageQueue) {
+    return this.itemsenseService.makeRequest(this.queueModel, MessageQueue.requestTypes.CONFIGURE, messageQueue);
   }
 
 }